requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
websockets==12.0
redis==5.0.4
supabase==2.5.0
pydantic==2.7.2
python-dotenv==1.0.1
sentence-transformers==3.0.0
requests==2.32.2

httpx==0.27.0
beautifulsoup4==4.12.3
transformers==4.41.2
torch
